Ajuste simples no tipo das variaveis.

+++ Simple&FastCalc/CalcCommands.cs
@@ -3,22 +3,22 @@ using Cocona;
 public class CalcCommands
 {
     [Command("add")]
-    public void Add(int a, int b)
+    public void Add(decimal a, decimal b)
     {
         Console.WriteLine($"Result: {a + b}");
     }
     [Command("subtc")]
-    public void Subtract(int a, int b)
+    public void Subtract(decimal a, decimal b)
     {
         Console.WriteLine($"Result: {a - b}");
     }
     [Command("mult")]
-    public void Multiply(int a, int b)
+    public void Multiply(decimal a, decimal b)
     {
         Console.WriteLine($"Result: {a * b}");
     }
     [Command("div")]
-    public void Division(int a, int b)
+    public void Division(decimal a, decimal b)
     {
         if (b == 0)
         {
@@ -30,7 +30,7 @@ public class CalcCommands
         }
     }
     [Command("rdiv")]
-    public void LeftDivision(int a, int b)
+    public void LeftDivision(decimal a, decimal b)
     {
         if (b == 0)
         {
